chore: refactor all mutations, getters and actions into their own files
