+++ src/store/actions/createChannel.ts
@@ -1,0 +1,43 @@
+import { v4 as uuidv4 } from "uuid";
+import { Commit } from "vuex";
+import { createChannel } from "@/core/methods/createChannel";
+
+import { MembraneType } from "@/store";
+
+export interface Context {
+  commit: Commit;
+  getters: any;
+}
+
+export interface Payload {
+  communityId: string;
+  name: string;
+}
+
+export default async (
+  { commit, getters }: Context,
+  { communityId, name }: Payload
+): Promise<void> => {
+  const community = getters.getCommunity(communityId);
+  const uid = uuidv4().toString();
+  const channel = await createChannel(
+    name,
+    "",
+    uid,
+    community.perspective,
+    community.linkLanguageAddress,
+    community.expressionLanguages,
+    MembraneType.Inherited,
+    community.typedExpressionLanguages
+  );
+
+  commit("addChannel", {
+    communityId: community.perspective,
+    channel,
+  });
+
+  // Need error handling (try catch in the awaits over)
+  return new Promise((resolve, reject) => {
+    resolve();
+  });
+};
