+++ src/store/actions/index.ts
@@ -1,0 +1,11 @@
+import createChannel from "./createChannel";
+import createCommunity from "./createCommunity";
+import joinCommunity from "./joinCommunity";
+import getPerspectiveChannelsAndMetadata from "./getPerspectiveChannelsAndMetadata";
+
+export default {
+  createChannel,
+  createCommunity,
+  joinCommunity,
+  getPerspectiveChannelsAndMetadata,
+};
